// src/api/generated/model/grantResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GrantResponseScopeType } from './grantResponseScopeType';

export interface GrantResponse {
  id?: string;
  userId?: string;
  scopeType?: GrantResponseScopeType;
  namespace?: string;
  datasourceId?: string;
  datasourceName?: string;
  readOnly?: boolean;
  createdAt?: string;
}